Add files via upload
Looking for help. The assignment is asking that the user inputs the number of trips. It needs to be between 1 and 13. Enters miles driven and has to be between 10 and 300. Enters gallons used between 1 to 18. again, all inputs need to be verified. Then we need to compute the MPG for each trip. All of this needs to be printed in tabular output. So, all the trips inputted under one section, miles, gallons, and mpg.